Many small fixes to lualine ... [nvim]
Position section in lualine has icons for columns and rows
Better padding for position section.
Add some extension to lualine (though it is not clear if they work correctly)
Fix padding and truncation of sections when the terminal is too narrow.
Remove custom function for diff
Use standard symbols for diff
No icon for diagnostics.